# docs/api-spec.yaml
openapi: 3.0.0
info:
  title: Prompt Testing Lab API
  version: 1.0.0
  description: |
    Enterprise-grade A/B testing platform for LLM prompts. 
    
    This API enables organizations to systematically test, compare, and optimize their AI prompts across multiple providers, reducing costs while improving output quality and consistency.
    
    ## Authentication
    All endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limits
    - Free tier: 1,000 requests/hour
    - Pro tier: 10,000 requests/hour
    - Enterprise: Custom limits
    
    ## Error Handling
    The API uses conventional HTTP response codes to indicate success or failure. Error responses include detailed error messages and error codes for programmatic handling.
    
  contact:
    name: Prompt Testing Lab Support
    url: https://promptlab.ai/support
    email: support@promptlab.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.promptlab.ai/v1
    description: Production server
  - url: https://staging-api.promptlab.ai/v1
    description: Staging server
  - url: http://localhost:8001/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "SecurePass123!"
                name:
                  type: string
                  minLength: 2
                  example: "John Doe"
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Authenticate user
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        200:
          description: Successful authentication
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Refresh authentication token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
...
  /projects/{projectId}/prompts:
    get:
      summary: List project prompts
      tags:
        - Prompts
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: tag
          in: query
          description: Filter by tag
          schema:
            type: string
        - name: version
          in: query
          description: Filter by version
          schema:
            type: integer
      responses:
        200:
          description: Prompts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prompt'

    post:
      summary: Create new prompt
      tags:
        - Prompts
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptCreate'
      responses:
        201:
          description: Prompt created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Prompt'
  /prompts/{promptId}:
    get:
      summary: Get prompt details
      tags:
        - Prompts
      parameters:
        - $ref: '#/components/parameters/PromptIdParam'
      responses:
        200:
          description: Prompt retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PromptDetails'

    put:
      summary: Update prompt
      tags:
        - Prompts
      parameters:
        - $ref: '#/components/parameters/PromptIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptUpdate'
      responses:
        200:
          description: Prompt updated successfully
  /prompts/{promptId}/versions:
    get:
      summary: Get prompt version history
      tags:
        - Prompts
      parameters:
        - $ref: '#/components/parameters/PromptIdParam'
      responses:
        200:
          description: Version history retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromptVersion'

    post:
      summary: Create new prompt version
      tags:
        - Prompts
      parameters:
        - $ref: '#/components/parameters/PromptIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptVersionCreate'
      responses:
        201:
          description: New version created

  # Test Execution
...
  /projects/{projectId}/analytics:
    get:
      summary: Get project analytics
      tags:
        - Analytics
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: period
          in: query
          description: Time period for analytics
          schema:
            type: string
            enum: [1d, 7d, 30d, 90d, 1y]
            default: 7d
        - name: groupBy
          in: query
          description: Group results by dimension
          schema:
            type: string
            enum: [model, prompt, day, hour]
      responses:
        200:
          description: Analytics data retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AnalyticsData'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      description: Unique identifier for the project
      schema:
        type: string
        format: uuid
    PromptIdParam:
      name: promptId
      in: path
      required: true
      description: Unique identifier for the prompt
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Refresh token for getting new access tokens
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, user, viewer]
        avatar:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time

    # Project Schemas
...
    Prompt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        template:
          type: string
        version:
          type: integer
        parentPromptId:
          type: string
          format: uuid
        modelConfig:
          $ref: '#/components/schemas/ModelConfig'
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PromptDetails:
      allOf:
        - $ref: '#/components/schemas/Prompt'
        - type: object
          properties:
            versions:
              type: array
              items:
                $ref: '#/components/schemas/PromptVersion'
            recentTestRuns:
              type: array
              items:
                $ref: '#/components/schemas/TestRun'
    PromptVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        template:
          type: string
        changes:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    PromptCreate:
      type: object
      required:
        - name
        - template
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        template:
          type: string
          minLength: 1
        modelConfig:
          $ref: '#/components/schemas/ModelConfig'
        tags:
          type: array
          items:
            type: string
    PromptUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        template:
          type: string
        modelConfig:
          $ref: '#/components/schemas/ModelConfig'
        tags:
          type: array
          items:
            type: string
    PromptVersionCreate:
      type: object
      required:
        - template
      properties:
        template:
          type: string
        changes:
          type: string
        modelConfig:
          $ref: '#/components/schemas/ModelConfig'

    # Model Configuration Schema
    ModelConfig:
      type: object
      required:
        - provider
        - modelName
      properties:
        provider:
          type: string
          enum: [openai, anthropic, groq, local]
        modelName:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        maxTokens:
          type: integer
          minimum: 1
          maximum: 32000
        topP:
          type: number
          minimum: 0
          maximum: 1
        frequencyPenalty:
          type: number
          minimum: -2
          maximum: 2
        presencePenalty:
          type: number
          minimum: -2
          maximum: 2
        systemMessage:
          type: string
        customEndpoint:
          type: string
          format: uri

    # Test Execution Schemas
    TestRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/TestRunStatus'
        progress:
          $ref: '#/components/schemas/TestProgress'
        config:
          type: object
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    TestRunStatus:
      type: string
      enum: [pending, queued, running, completed, failed, cancelled]
...
    TestProgress:
      type: object
      properties:
        completed:
          type: integer
        total:
          type: integer
        percentage:
          type: number
          minimum: 0
          maximum: 100
        currentPrompt:
          type: string
        estimatedTimeRemaining:
          type: integer
          description: Estimated seconds remaining
...
    AnalyticsData:
      type: object
      properties:
        period:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        metrics:
          type: object
          properties:
            totalTestRuns:
              type: integer
            totalPrompts:
              type: integer
            avgExecutionTime:
              type: number
            totalCost:
              type: number
            successRate:
              type: number
        trends:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              testRuns:
                type: integer
              cost:
                type: number
              avgLatency:
                type: number
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...